requirements.txt
azure-cognitiveservices-speech
python-dotenv
Pillow
